feat #85 : ResponseDTO 솔로 및 크루로 나누기, 노션 명세서에 맞추서 순서 수정
