+++ src/main/java/com/umc/yourun/controller/SoloChallengeRestController.java
@@ -3,7 +3,8 @@ package com.umc.yourun.controller;
 import com.umc.yourun.apiPayload.ApiResponse;
 import com.umc.yourun.config.exception.ErrorResponse;
 import com.umc.yourun.dto.challenge.ChallengeRequest;
-import com.umc.yourun.dto.challenge.ChallengeResponse;
+import com.umc.yourun.dto.challenge.CrewChallengeResponse;
+import com.umc.yourun.dto.challenge.SoloChallengeResponse;
 import com.umc.yourun.service.challenge.SoloChallengeService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -29,10 +30,10 @@ public class SoloChallengeRestController {
                     content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     })
     @PostMapping("")
-    public ApiResponse<ChallengeResponse.SoloChallengeCreateRes> createSoloChallenge(
+    public ApiResponse<SoloChallengeResponse.SoloChallengeCreateRes> createSoloChallenge(
             @RequestHeader(value = "Authorization") String accessToken,
             @RequestBody @Valid ChallengeRequest.CreateSoloChallengeReq request) {
-        ChallengeResponse.SoloChallengeCreateRes response = soloChallengeService.createSoloChallenge(request, accessToken);
+        SoloChallengeResponse.SoloChallengeCreateRes response = soloChallengeService.createSoloChallenge(request, accessToken);
         return ApiResponse.success("솔로 챌린지가 생성되었습니다.", response);
     }
 
@@ -43,65 +44,65 @@ public class SoloChallengeRestController {
                     content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     })
     @GetMapping("/pending")
-    public ApiResponse<ChallengeResponse.SoloChallenge> getPendingSoloChallenges(
+    public ApiResponse<SoloChallengeResponse.SoloChallengeRes> getPendingSoloChallenges(
             @RequestHeader(value = "Authorization") String accessToken
     ) {
-        ChallengeResponse.SoloChallenge result = soloChallengeService.getPendingSoloChallenges(accessToken);
+        SoloChallengeResponse.SoloChallengeRes result = soloChallengeService.getPendingSoloChallenges(accessToken);
         return ApiResponse.success("대기 중인 솔로 챌린지 목록입니다.", result);
     }
 
-    @Operation(summary = "SOLO_CHALLENGE_API_03 : 솔로 챌린지 참여", description = "대기 중인 솔로 챌린지에 참여합니다.")
-    @ApiResponses(value = {
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "챌린지 참여 성공"),
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "400", description = "잘못된 요청",
-                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
-    })
-    @PostMapping("/{challengeId}/join")
-    public ApiResponse<ChallengeResponse.SoloChallengeMateRes> joinSoloChallenge(
-            @RequestHeader(value = "Authorization") String accessToken,
-            @PathVariable Long challengeId) {
-        ChallengeResponse.SoloChallengeMateRes response = soloChallengeService.joinSoloChallenge(challengeId, accessToken);
-        return ApiResponse.success("솔로 챌린지 참여가 완료되었습니다.", response);
-    }
-
-    @Operation(summary = "SOLO_CHALLENGE_API_04 : 솔로 챌린지 상세 정보 조회", description = "대기 중인 솔로 챌린지 상세 조회 화면입니다.")
+    @Operation(summary = "SOLO_CHALLENGE_API_03 : 솔로 챌린지 상세 정보 조회", description = "대기 중인 솔로 챌린지 상세 조회 화면입니다.")
     @ApiResponses(value = {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "챌린지 참여 성공"),
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "400", description = "잘못된 요청",
                     content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     })
     @GetMapping("/pending/{challengeId}")
-    public ApiResponse<ChallengeResponse.SoloChallengeDetailRes> getSoloChallengeDetail(
+    public ApiResponse<SoloChallengeResponse.SoloChallengeDetailRes> getSoloChallengeDetail(
             @RequestHeader(value = "Authorization") String accessToken,
             @PathVariable Long challengeId) {
-        ChallengeResponse.SoloChallengeDetailRes response = soloChallengeService.getSoloChallengeDetail(challengeId, accessToken);
+        SoloChallengeResponse.SoloChallengeDetailRes response = soloChallengeService.getSoloChallengeDetail(challengeId, accessToken);
         return ApiResponse.success("솔로 챌린지 상세 조회 정보입니다.", response);
     }
 
-    @Operation(summary = "SOLO_CHALLENGE_API_05 : 솔로 챌린지 매칭 조회", description = "매칭된 솔로 챌린지 조회 화면입니다.")
+    @Operation(summary = "SOLO_CHALLENGE_API_04 : 솔로 챌린지 매칭 조회", description = "매칭된 솔로 챌린지 조회 화면입니다.")
     @ApiResponses(value = {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "챌린지 참여 성공"),
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "400", description = "잘못된 요청",
                     content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     })
     @GetMapping("/matching")
-    public ApiResponse<ChallengeResponse.SoloChallengeMatchingRes> getSoloMatch (
+    public ApiResponse<SoloChallengeResponse.SoloChallengeMatchingRes> getSoloMatch (
             @RequestHeader(value = "Authorization") String accessToken) {
-        ChallengeResponse.SoloChallengeMatchingRes response = soloChallengeService.getSoloChallengeMatching(accessToken);
+        SoloChallengeResponse.SoloChallengeMatchingRes response = soloChallengeService.getSoloChallengeMatching(accessToken);
         return ApiResponse.success("솔로 챌린지 매칭 정보입니다.", response);
     }
 
-    @Operation(summary = "SOLO_CHALLENGE_API_06 : 솔로 챌린지 일자별 진행도 조회", description = "솔로 챌린지의 일자별 조회 화면입니다.")
+    @Operation(summary = "SOLO_CHALLENGE_API_05 : 솔로 챌린지 일자별 진행도 조회", description = "솔로 챌린지의 일자별 조회 화면입니다.")
     @ApiResponses(value = {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "챌린지 참여 성공"),
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "400", description = "잘못된 요청",
                     content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     })
     @GetMapping("/progress")
-    public ApiResponse<ChallengeResponse.SoloChallengeProgressRes> getSoloProgress (
+    public ApiResponse<SoloChallengeResponse.SoloChallengeProgressRes> getSoloProgress (
             @RequestHeader(value = "Authorization") String accessToken) {
-        ChallengeResponse.SoloChallengeProgressRes response = soloChallengeService.getSoloChallengeProgress(accessToken);
+        SoloChallengeResponse.SoloChallengeProgressRes response = soloChallengeService.getSoloChallengeProgress(accessToken);
         return ApiResponse.success("솔로 챌린지 일자별 진행도 조회 정보입니다.", response);
+    }
+
+    @Operation(summary = "SOLO_CHALLENGE_API_07 : 솔로 챌린지 참여", description = "대기 중인 솔로 챌린지에 참여합니다.")
+    @ApiResponses(value = {
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "챌린지 참여 성공"),
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "400", description = "잘못된 요청",
+                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+    })
+    @PostMapping("/{challengeId}/join")
+    public ApiResponse<SoloChallengeResponse.SoloChallengeMateRes> joinSoloChallenge(
+            @RequestHeader(value = "Authorization") String accessToken,
+            @PathVariable Long challengeId) {
+        SoloChallengeResponse.SoloChallengeMateRes response = soloChallengeService.joinSoloChallenge(challengeId, accessToken);
+        return ApiResponse.success("솔로 챌린지 참여가 완료되었습니다.", response);
     }
 
 }
