+++ src/main/java/com/umc/yourun/controller/UserController.java
@@ -1,9 +1,9 @@
 package com.umc.yourun.controller;
 
 import com.umc.yourun.apiPayload.ApiResponse;
-import com.umc.yourun.config.exception.ErrorCode;
 import com.umc.yourun.config.exception.ErrorResponse;
-import com.umc.yourun.dto.challenge.ChallengeResponse;
+import com.umc.yourun.dto.challenge.CrewChallengeResponse;
+import com.umc.yourun.dto.challenge.SoloChallengeResponse;
 import com.umc.yourun.dto.user.UserRequestDTO;
 import com.umc.yourun.service.UserService;
 import com.umc.yourun.service.challenge.SoloChallengeService;
@@ -12,7 +12,6 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.Email;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
@@ -61,9 +60,9 @@ public class UserController {
                     content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     })
     @GetMapping("/home/challenges")
-    public ApiResponse<ChallengeResponse.HomeChallengeRes> getUserChallenges(
+    public ApiResponse<SoloChallengeResponse.HomeChallengeRes> getUserChallenges(
             @RequestHeader(value = "Authorization") String accessToken) {
-        ChallengeResponse.HomeChallengeRes response = soloChallengeService.getUserChallenges(accessToken);
+        SoloChallengeResponse.HomeChallengeRes response = soloChallengeService.getUserChallenges(accessToken);
         return ApiResponse.success("홈 화면 : 사용자와 관련된 챌린지 정보입니다.", response);
     }
