+++ src/main/java/com/umc/yourun/service/challenge/SoloChallengeService.java
@@ -13,9 +13,9 @@ import com.umc.yourun.domain.enums.*;
 import com.umc.yourun.domain.mapping.UserCrewChallenge;
 import com.umc.yourun.domain.mapping.UserSoloChallenge;
 import com.umc.yourun.dto.challenge.ChallengeRequest;
-import com.umc.yourun.dto.challenge.ChallengeResponse;
+import com.umc.yourun.dto.challenge.CrewChallengeResponse;
+import com.umc.yourun.dto.challenge.SoloChallengeResponse;
 import com.umc.yourun.repository.*;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -48,7 +48,7 @@ public class SoloChallengeService {
 
     // 1. 솔로 챌린지 생성
     @Transactional
-    public ChallengeResponse.SoloChallengeCreateRes createSoloChallenge(ChallengeRequest.CreateSoloChallengeReq request, String accessToken) {
+    public SoloChallengeResponse.SoloChallengeCreateRes createSoloChallenge(ChallengeRequest.CreateSoloChallengeReq request, String accessToken) {
 
         // 유저 조회
         User user = jwtTokenProvider.getUserByToken(accessToken);
@@ -88,25 +88,25 @@ public class SoloChallengeService {
         UserSoloChallenge userSoloChallenge = ChallengeConverter.toUserSoloChallenge(user, savedSoloChallenge, true);
         userSoloChallengeRepository.save(userSoloChallenge);
 
-        return new ChallengeResponse.SoloChallengeCreateRes(savedSoloChallenge.getId(),
+        return new SoloChallengeResponse.SoloChallengeCreateRes(savedSoloChallenge.getId(),
                 formatDateTime(startDateTime), formatDateTime(endDateTime),
                 savedSoloChallenge.getChallengePeriod().getDays(), user.getTendency());
     }
 
     // 2. PENDING 상태인 솔로 챌린지 조회 (화면)
     @Transactional(readOnly = true)
-    public ChallengeResponse.SoloChallenge getPendingSoloChallenges(String accessToken) {
+    public SoloChallengeResponse.SoloChallengeRes getPendingSoloChallenges(String accessToken) {
         // 유저 조회
         User user = jwtTokenProvider.getUserByToken(accessToken);
-        List<ChallengeResponse.SoloChallengeRes> pendingSoloChallenges = getPendingSoloChallenge();
+        List<SoloChallengeResponse.SoloChallenge> pendingSoloChallenges = getPendingSoloChallenge();
 
-        return new ChallengeResponse.SoloChallenge(user.getId(), user.getTendency(),
+        return new SoloChallengeResponse.SoloChallengeRes(user.getId(), user.getTendency(),
                 user.getCrewReward(), user.getPersonalReward(), pendingSoloChallenges);
     }
 
     // 2-1. PENDING 상태인 솔로 챌린지
     @Transactional(readOnly = true)
-    public List<ChallengeResponse.SoloChallengeRes> getPendingSoloChallenge() {
+    public List<SoloChallengeResponse.SoloChallenge> getPendingSoloChallenge() {
 
 
         List<SoloChallenge> pendingChallenges = soloChallengeRepository.findRandomPendingChallenges(5);
@@ -137,7 +137,7 @@ public class SoloChallengeService {
                         default -> 0;
                     };
 
-                    return new ChallengeResponse.SoloChallengeRes(
+                    return new SoloChallengeResponse.SoloChallenge(
                             challenge.getId(),
                             challenge.getChallengeDistance().getDistance(),
                             challenge.getChallengePeriod().getDays(),
@@ -152,7 +152,7 @@ public class SoloChallengeService {
 
     // 3. 솔로 챌린지 상세 조회
     @Transactional(readOnly = true)
-    public ChallengeResponse.SoloChallengeDetailRes getSoloChallengeDetail(Long challengeId, String accessToken) {
+    public SoloChallengeResponse.SoloChallengeDetailRes getSoloChallengeDetail(Long challengeId, String accessToken) {
 
         // 유저 조회
         User user = jwtTokenProvider.getUserByToken(accessToken);
@@ -186,7 +186,7 @@ public class SoloChallengeService {
 
         int countDay = calculateCountDay(user.getCreatedAt().toLocalDate());
 
-        return new ChallengeResponse.SoloChallengeDetailRes(formatDateTime(soloChallenge.get().getStartDate()),
+        return new SoloChallengeResponse.SoloChallengeDetailRes(formatDateTime(soloChallenge.get().getStartDate()),
                 formatDateTime(soloChallenge.get().getEndDate()), soloChallenge.get().getChallengeDistance().getDistance(),
                 soloChallenge.get().getChallengePeriod().getDays(), nickname, hashtags, creatorUser.getTendency(),
                 reward, countDay);
@@ -194,7 +194,7 @@ public class SoloChallengeService {
 
     // 4. 솔로 챌린지 매칭 화면
     @Transactional(readOnly = true)
-    public ChallengeResponse.SoloChallengeMatchingRes getSoloChallengeMatching(String accessToken) {
+    public SoloChallengeResponse.SoloChallengeMatchingRes getSoloChallengeMatching(String accessToken) {
 
         // 유저 조회
         User user = jwtTokenProvider.getUserByToken(accessToken);
@@ -226,7 +226,7 @@ public class SoloChallengeService {
         int userCountDay = calculateCountDay(user.getCreatedAt().toLocalDate());
         int challengeMateCountDay = calculateCountDay(challengeMate.get().getCreatedAt().toLocalDate());
 
-        return new ChallengeResponse.SoloChallengeMatchingRes(soloChallenge.getChallengePeriod().getDays(),
+        return new SoloChallengeResponse.SoloChallengeMatchingRes(soloChallenge.getChallengePeriod().getDays(),
                 soloChallenge.getChallengeDistance().getDistance(), user.getTendency(),
                 user.getNickname(), userCountDay, userHashtags, challengeMate.get().getTendency(),
                 challengeMate.get().getNickname(), challengeMateCountDay, challengeHashtags);
@@ -234,7 +234,7 @@ public class SoloChallengeService {
     }
 
     // 5. 솔로 챌린지 일자별 진행도
-    public ChallengeResponse.SoloChallengeProgressRes getSoloChallengeProgress(String accessToken) {
+    public SoloChallengeResponse.SoloChallengeProgressRes getSoloChallengeProgress(String accessToken) {
 
         // 유저 조회
         User user = jwtTokenProvider.getUserByToken(accessToken);
@@ -252,15 +252,15 @@ public class SoloChallengeService {
         Optional<User> challengeMate = userRepository.findById(mateId);
 
         // 일자별 러닝 데이터 조회
-        List<ChallengeResponse.DayResult> dayResults = getDayResults(user.getId(), soloChallenge);
+        List<SoloChallengeResponse.DayResult> dayResults = getDayResults(user.getId(), soloChallenge);
 
         // 챌린지 메이트의 거리 및 정보 조회
-        ChallengeResponse.ChallengeMateInfo mateInfo = getChallengeMateInfo(challengeMate.get(), soloChallenge);
+        SoloChallengeResponse.ChallengeMateInfo mateInfo = getChallengeMateInfo(challengeMate.get(), soloChallenge);
 
         // 금일 성공 여부
         boolean isSuccess = checkTodaySuccess(user.getId(), soloChallenge);
 
-        return new ChallengeResponse.SoloChallengeProgressRes(soloChallenge.getChallengePeriod().getDays(),
+        return new SoloChallengeResponse.SoloChallengeProgressRes(soloChallenge.getChallengePeriod().getDays(),
                 soloChallenge.getChallengeDistance().getDistance(),
                 calculateDayCount(soloChallenge.getStartDate()),
                 formatDateTime(LocalDateTime.now()),
@@ -273,7 +273,7 @@ public class SoloChallengeService {
     }
 
     // 5-1. 일자별 달린 거리
-    private List<ChallengeResponse.DayResult> getDayResults(Long userId, SoloChallenge challenge) {
+    private List<SoloChallengeResponse.DayResult> getDayResults(Long userId, SoloChallenge challenge) {
         LocalDateTime startDate = challenge.getStartDate();
         LocalDateTime now = parseDateTime(formatDateTime(LocalDateTime.now())); // 시간과 분만 비교할 수 있도록 포멧팅
 
@@ -309,13 +309,13 @@ public class SoloChallengeService {
         }
 
         return dailyDistances.entrySet().stream()
-                .map(entry -> new ChallengeResponse.DayResult(entry.getKey(), entry.getValue()))
-                .sorted(Comparator.comparing(ChallengeResponse.DayResult::day))
+                .map(entry -> new SoloChallengeResponse.DayResult(entry.getKey(), entry.getValue()))
+                .sorted(Comparator.comparing(SoloChallengeResponse.DayResult::day))
                 .collect(Collectors.toList());
     }
 
     // 5-2. 챌린지 메이트 달린 거리 및 정보
-    private ChallengeResponse.ChallengeMateInfo getChallengeMateInfo(User mate, SoloChallenge challenge) {
+    private SoloChallengeResponse.ChallengeMateInfo getChallengeMateInfo(User mate, SoloChallenge challenge) {
         LocalDateTime now = parseDateTime(formatDateTime(LocalDateTime.now())); // 시간과 분만 비교할 수 있도록 포멧팅
         int currentDay = calculateDayFromStart(challenge.getStartDate(), now);
 
@@ -349,7 +349,7 @@ public class SoloChallengeService {
                     int successDays = countSuccessDaysBeforeFailure(mate.getId(), challenge, day - 1);
                     double totalDistance = calculateTotalDistanceBeforeFailure(mate.getId(), challenge, day - 1);
 
-                    return new ChallengeResponse.ChallengeMateInfo(
+                    return new SoloChallengeResponse.ChallengeMateInfo(
                             mate.getNickname(),
                             mate.getTendency(),
                             successDays,
@@ -388,7 +388,7 @@ public class SoloChallengeService {
             successDays++;
         }
 
-        return new ChallengeResponse.ChallengeMateInfo(
+        return new SoloChallengeResponse.ChallengeMateInfo(
                 mate.getNickname(),
                 mate.getTendency(),
                 successDays,
@@ -399,7 +399,7 @@ public class SoloChallengeService {
 
     // 7. 솔로 챌린지에 참여하기
     @Transactional
-    public ChallengeResponse.SoloChallengeMateRes joinSoloChallenge(Long challengeId, String accessToken) {
+    public SoloChallengeResponse.SoloChallengeMateRes joinSoloChallenge(Long challengeId, String accessToken) {
 
         // 유저 조회
         User user = jwtTokenProvider.getUserByToken(accessToken);
@@ -445,12 +445,12 @@ public class SoloChallengeService {
         // 8. 챌린지 상태 업데이트
         soloChallenge.updateStatus(ChallengeStatus.IN_PROGRESS);
 
-        return new ChallengeResponse.SoloChallengeMateRes(challengeId, creatorChallenge.getUser().getId());
+        return new SoloChallengeResponse.SoloChallengeMateRes(challengeId, creatorChallenge.getUser().getId());
     }
 
     // 홈 화면에서 유저의 챌린지 관련 화면 조회
     @org.springframework.transaction.annotation.Transactional(readOnly = true)
-    public ChallengeResponse.HomeChallengeRes getUserChallenges(String accessToken) {
+    public SoloChallengeResponse.HomeChallengeRes getUserChallenges(String accessToken) {
 
         // 유저 조회
         User user = jwtTokenProvider.getUserByToken(accessToken);
@@ -469,7 +469,7 @@ public class SoloChallengeService {
                         List.of(ChallengeStatus.PENDING, ChallengeStatus.IN_PROGRESS));
 
         // 응답 DTO 생성
-        ChallengeResponse.UserSoloChallengeInfo soloInfo = null;
+        SoloChallengeResponse.UserSoloChallengeInfo soloInfo = null;
         if (userSoloChallenge != null) {
             SoloChallenge challenge = userSoloChallenge.getSoloChallenge();
             int soloCountDay = calculateCountDay(challenge.getStartDate().toLocalDate());
@@ -499,7 +499,7 @@ public class SoloChallengeService {
             soloInfo = ChallengeConverter.toUserSoloChallengeInfo(challenge, user, mateId, mateNickname, tendency, soloCountDay, formatDateTime(soloStartDate));
         }
 
-        ChallengeResponse.UserCrewChallengeInfo crewInfo = null;
+        SoloChallengeResponse.UserCrewChallengeInfo crewInfo = null;
         LocalDateTime crewStartDate = userCrewChallenge.getCrewChallenge().getStartDate();
         if (userCrewChallenge != null) {
             CrewChallenge challenge = userCrewChallenge.getCrewChallenge();
@@ -508,13 +508,13 @@ public class SoloChallengeService {
             // 크루원 수 확인
             long memberCount = userCrewChallengeRepository.countByCrewChallengeId(challenge.getId());
             if (memberCount == 4) {  // 4명이 모인 크루만 응답
-                List<ChallengeResponse.MemberTendencyInfo> participantInfos = crewChallengeService.getMemberTendencyInfos(challenge.getId());
+                List<CrewChallengeResponse.MemberTendencyInfo> participantInfos = crewChallengeService.getMemberTendencyInfos(challenge.getId());
 
                 crewInfo = ChallengeConverter.toUserCrewChallengeInfo(challenge, participantInfos, crewCountDay, formatDateTime(crewStartDate));
             }
         }
 
-        return new ChallengeResponse.HomeChallengeRes(soloInfo, crewInfo);
+        return new SoloChallengeResponse.HomeChallengeRes(soloInfo, crewInfo);
     }
